requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==4.0.4
djangorestframework==3.15.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
gdown==5.1.0
google-pasta==0.2.0
grpcio==1.63.0
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opennsfw2==0.13.7
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.31.0
rich==13.7.1
scikit-image==0.22.0
scipy==1.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.5.3
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
whitenoise==6.6.0
wrapt==1.16.0
zipp==3.18.1
